Restructure project to a single-file format
