--- YTDownloader/__init__.py
@@ -1,10 +1,0 @@
-"""A program to download any YouTube video or playlist."""
-from __future__ import annotations
-
-__version__: Final[str] = "1.5.1"
-__title__: Final[str] = "YTDownloader"
-__author__: Final[str] = "realshouzy"
-__license__: Final[str] = "MIT"
-__copyright__: Final[str] = "Copyright (c) 2022-present realshouzy"
-
-from typing import Final
