--- YTDownloader/__main__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-"""Run module as file."""
-from __future__ import annotations
-
-from YTDownloader.main import main
-
-if __name__ == "__main__":
-    raise SystemExit(main())
